% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_dat}
\alias{all_dat}
\title{All CPI Data, by Major Group}
\format{A list with nine elements, one for each major group, with endpoints:
\describe{
\item{CPI}{cu.data.1.AllItems}
\item{Food and Bev}{cu.data.11.USFoodBeverage}
\item{Housing}{cu.data.12.USHousing}
\item{Apparel}{cu.data.13.USApparel}
\item{Transportation}{cu.data.14.USTransportation}
\item{Medical}{cu.data.15.USMedical}
\item{Recreation}{cu.data.16.USRecreation}
\item{Eduction and Communication}{cu.data.17.USEducationAndCommunication}
\item{Other Good and Services}{cu.data.18.USOtherGoodsAndServices}
}}
\source{
\url{https://download.bls.gov/pub/time.series/cu/{...}}
}
\usage{
all_dat
}
\description{
All CPI Data, by Major Group
}
\keyword{datasets}
