% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_dat}
\alias{area_dat}
\title{Geographic Region (Area) Codes}
\format{A data frame with 58 rows and three columns:
\describe{
\item{area_code}{BLS four digit geographic region code}
\item{area_name}{Name of the region}
\item{display_level}{Level in categorical hierarchy}
}}
\source{
\url{https://download.bls.gov/pub/time.series/cu/cu.area}
}
\usage{
area_dat
}
\description{
Dataset with geographic codes.
}
\keyword{datasets}
