% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\name{assign_mosquito_status}
\alias{assign_mosquito_status}
\title{Given a number of bites, assign them to classes of mosquitoes.}
\usage{
assign_mosquito_status(bites_df, M, Y, Z)
}
\arguments{
\item{bites_df}{has \code{human_level} and \code{times} for each bite}

\item{M}{is number of adult females}

\item{Y}{is number of exposed adult females}

\item{Z}{is number of infectious adult females}
}
\value{
adds \code{infect_mosquito} and \code{infect_human} to the data frame.
}
\description{
Given a number of bites, assign them to classes of mosquitoes.
}
