% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_linear.R
\name{bites_at_location}
\alias{bites_at_location}
\title{Assign bites to humans present at location.
Events has columns (ID, Level, Time, Location, Event).
Events include enter 1, leave -1, and change infectiousness level 2.
Bites has columns (Location, Bite, Time).
Returned bites have an ID for the human, that can be -1 for not-a-human,
and a Level, for the infectiousness of the human bitten.}
\usage{
bites_at_location(events, bites)
}
\description{
Assign bites to humans present at location.
Events has columns (ID, Level, Time, Location, Event).
Events include enter 1, leave -1, and change infectiousness level 2.
Bites has columns (Location, Bite, Time).
Returned bites have an ID for the human, that can be -1 for not-a-human,
and a Level, for the infectiousness of the human bitten.
}
