% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\name{bld_bite_outcomes}
\alias{bld_bite_outcomes}
\title{Iterate over locations and calculate bloodmeal for each location
location_events has columns \code{ID}, \code{Level}, \code{Time}, \code{Location}, \code{Event}.
Level is 0 for uninfectious, 1 for infectious. Event is 1, -1 for enter
and leave, 2 for change in infectious status.}
\usage{
bld_bite_outcomes(location_events, bites)
}
\description{
Iterate over locations and calculate bloodmeal for each location
location_events has columns \code{ID}, \code{Level}, \code{Time}, \code{Location}, \code{Event}.
Level is 0 for uninfectious, 1 for infectious. Event is 1, -1 for enter
and leave, 2 for change in infectious status.
}
