% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\name{bld_convert_to_enter_events}
\alias{bld_convert_to_enter_events}
\title{Takes location movements and turns them into a sequence of entering
leaving events for each location.}
\usage{
bld_convert_to_enter_events(location_changes_dt)
}
\description{
Takes location movements and turns them into a sequence of entering
leaving events for each location.
}
