% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\name{bld_convert_to_source_destination}
\alias{bld_convert_to_source_destination}
\title{Converts from incoming wide format data table to a long one.}
\usage{
bld_convert_to_source_destination(location_dt)
}
\arguments{
\item{location_dt}{Data table with one individual per row and two columns
for each movement. One column for when, one for where.}
}
\value{
Output data table is a sequence of movements, so columns are
person, time, source, destination.
}
\description{
Converts from incoming wide format data table to a long one.
}
