% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\name{bld_location_next_state}
\alias{bld_location_next_state}
\title{Tells you the state of humans at a location at a given time.}
\usage{
bld_location_next_state(location_events, query_times, previous_state = NULL)
}
\arguments{
\item{location_events}{An event stream of movement and infection events for a single location.}

\item{query_times}{A length 2 numeric with beginning and ending times for the query.}

\item{previous_state}{The previously-returned value from this function.

This is for assigning outcomes to bites. Each call of this function processes
more of the event stream. It returns a data.table that has one entry for
each individual in the location at that time, with a Level for each of those
individuals. You then take the output from one call and pass it as
previous_state for the next call.}
}
\description{
Tells you the state of humans at a location at a given time.
}
