% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\name{bld_single_day}
\alias{bld_single_day}
\title{Assign bites to humans and mosquitoes for one day.}
\usage{
bld_single_day(
  day_idx,
  M_arr,
  Y_arr,
  Z_arr,
  biting_arr,
  dwell.lh,
  bite_weight,
  health_dt,
  params
)
}
\arguments{
\item{day_idx}{Index of day within the time step.}

\item{M_arr}{Matrix of location x day for total number of mosquitoes.}

\item{Y_arr}{Matrix of location x day for incubating mosquitoes.}

\item{Z_arr}{Matrix of location x day for infectious mosquitoes.}

\item{biting_arr}{Sample bites from this, location x day.}

\item{dwell.lh}{Total dwell time in each location by humans,
location x human.}

\item{bite_weight}{Biting weight of each human.}

\item{health_dt}{Health status from the human module.}

\item{params}{Module parameters.

This function samples bites in accordance with both the human
biting weight and the mosquito biting rate.}
}
\description{
Assign bites to humans and mosquitoes for one day.
}
