% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_constant.R
\name{bloodmeal_constant_module}
\alias{bloodmeal_constant_module}
\title{This bloodmeal is for testing. It sends out the same data each time.}
\usage{
bloodmeal_constant_module(parameters)
}
\arguments{
\item{parameters}{Configuration for the module. Must have
\code{duration_days}, \code{location_cnt}, \code{human_cnt}.}
}
\description{
This bloodmeal is for testing. It sends out the same data each time.
}
