% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\docType{class}
\name{bloodmeal_density}
\alias{bloodmeal_density}
\title{bloodmeal_density: Bloodmeal for mosquito modules that report mosquito density.}
\description{
The human module and location module both report events for individual humans.
The mosquito module reports changes in density of mosquitoes in each location
over time.
}
\details{
\itemize{
\item \link{bloodmeal_density_module}
\item \link{mash_step.bloodmeal_density}
\item \link{infects_human_path.bloodmeal_density}
\item \link{infects_mosquito_path.bloodmeal_density}
}
}
