% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_linear.R
\name{bloodmeal_linear_module}
\alias{bloodmeal_linear_module}
\title{Create a bloodmeal module that assigns the same weight to all humans.}
\usage{
bloodmeal_linear_module(parameters)
}
\arguments{
\item{parameters}{There are no parameters for this, so this is ignored.}
}
\value{
a module for bloodmeal
}
\description{
Create a bloodmeal module that assigns the same weight to all humans.
}
