% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_linear.R
\name{bloodmeal_process}
\alias{bloodmeal_process}
\title{Assign bites to people, the top-level function.}
\usage{
bloodmeal_process(health_dt, movement_dt, bites_dt)
}
\arguments{
\item{health_dt}{Health events from the protocol.}

\item{movement_dt}{Movement events from the protocol.}

\item{bites_dt}{Bite events from the protocol.}
}
\description{
Assign bites to people, the top-level function.
}
