% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{build_biting_state}
\alias{build_biting_state}
\title{Create a default state for mosquito-RM model.}
\usage{
build_biting_state(parameters)
}
\arguments{
\item{The}{internal parameter set.}
}
\description{
Create a default state for mosquito-RM model.
}
