% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_daily.R
\name{build_world}
\alias{build_world}
\title{Create parameters for a small model of mosquito bites.}
\usage{
build_world(locations, humans_per_location)
}
\arguments{
\item{locations}{The count of locations.}

\item{humans_per_location}{The count of humans per location, a single integer.}
}
\description{
Create parameters for a small model of mosquito bites.
}
