% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{check_continuous_setup}
\alias{check_continuous_setup}
\title{Check preconditions for the data structures.}
\usage{
check_continuous_setup(individuals, transitions)
}
\arguments{
\item{individuals}{a data frame of individuals, including columns for times.}

\item{transitions}{a list of transitions for enabling and firing times}
}
\description{
Check preconditions for the data structures.
}
