% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_equations.R
\name{cme_maker}
\alias{cme_maker}
\title{Use equations to make a function that calculates the matrix for the master equation.}
\usage{
cme_maker(equations)
}
\arguments{
\item{equations}{A list of formulas.}
}
\description{
Use equations to make a function that calculates the matrix for the master equation.
}
\seealso{
\code{\link{normalize_equations}}
}
