% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_linear.R
\name{combine_health_and_movement}
\alias{combine_health_and_movement}
\title{Makes a single time series of movement with health status}
\usage{
combine_health_and_movement(health_dt, movement_dt)
}
\description{
\preformatted{ID Level      Time Location Event
}

1:  1     0 0.0000000        2     1
2:  1     0 1.3734866        3     1
3:  1     0 1.3734866        2    -1
}
