% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{continuous_step}
\alias{continuous_step}
\title{Run until a specified time.}
\usage{
continuous_step(individuals, transitions)
}
\arguments{
\item{individuals}{a data frame of individuals, including columns for times.}

\item{transitions}{a list of transitions for enabling and firing times}

\item{end_time}{is a time beyond which nothing should fire}
}
\value{
a trajectory
}
\description{
Run until a specified time.
}
