% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert_to_r_movement}
\alias{convert_to_r_movement}
\title{Convert movement result to R data}
\usage{
convert_to_r_movement(movement, human)
}
\arguments{
\item{movement}{The movement module object}

\item{human}{An integer vector of which human positions to extract.}
}
\value{
Movement list with a "moves" list that has human locations.
}
\description{
The result of a movement object is in C++. This extracts
it into an R list. The result is added to the original
list that was passed in.
}
