% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CPI}
\alias{CPI}
\title{Consumer Price Index (CPI)}
\format{A data frame with 1263 rows, and four columns:
\describe{
\item{series_id}
\item{date}
\item{value}{CPI}
\item{lag1}{Change in CPI from previous observation.}
}}
\source{
\url{https://download.bls.gov/pub/time.series/cu/cu.data.1.AllItems}
}
\usage{
CPI
}
\description{
CPI: series id, date, value, and change
}
\keyword{datasets}
