% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\name{data_table_to_array}
\alias{data_table_to_array}
\title{Given a data table, generate an array form.}
\usage{
data_table_to_array(dt, row, col, value)
}
\arguments{
\item{dt}{the data table}

\item{row}{the name of the column to use as row values. Must be integers.}

\item{col}{name of column to use as column values. Must be integers.}

\item{value}{name of column to use as values.}
}
\value{
an array with size from min to max in row and column, assuming
they are integers that are in order.
}
\description{
Given a data table, generate an array form.
}
