% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_equations.R
\name{exact_exponential}
\alias{exact_exponential}
\title{Use equations to make a function that calculates the matrix for an exact time step.}
\usage{
exact_exponential(equations)
}
\arguments{
\item{equations}{A list of formulas.}
}
\description{
This exponentiates the matrix, giving \eqn{\exp(At)}.
}
\seealso{
\code{\link{normalize_equations}}
}
