% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_bites.R
\name{exponential_bite_module}
\alias{exponential_bite_module}
\title{Gives a constant rate of bites in a patch.}
\usage{
exponential_bite_module(parameters)
}
\arguments{
\item{parameters}{containing \code{people_cnt}, \code{bite_rate}, \code{duration}.}
}
\value{
A bite module.
}
\description{
Gives a constant rate of bites in a patch.
}
