% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{fire}
\alias{fire}
\title{Fire transition, changing the individual}
\usage{
fire(transition, individual, when, variables = NULL)
}
\arguments{
\item{transition}{A list of transitions}

\item{individual}{A list with individual properties}

\item{when}{A time}

\item{variables}{Any global simulation state}
}
\value{
a new list describing the individual
}
\description{
This function is for testing the transitions on data for a
single individual. It's a helper function. The running simulation
doesn't call this function.
}
\seealso{
\code{\link{is_enabled}}, \code{\link{when}}
}
