% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_flux.R
\name{flux_module}
\alias{flux_module}
\title{Creates a MASH module that runs Flux}
\usage{
flux_module(parameters)
}
\arguments{
\item{parameters}{A list or environment}
}
\value{
A simulation object, which is a list.
}
\description{
This module simulates movement according to a 'flux' model, where everybody
is indistinguishable and moves to a new patch based on where they are now.
}
\details{
The input \code{parameters} must have the following \emph{named} members:
\itemize{
\item \code{rate_matrix}: a matrix whose elements \eqn{\{f_{ij}\}} give the rate of movement
between those two patches. The diagonal must be all zeros.
\item \code{npatch}: number of patches/places
\item \code{location}: vector of everybody's current location
}
}
