% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flux_movement_init}
\alias{flux_movement_init}
\title{Initialize Flux Movement module}
\usage{
flux_movement_init(parameters, initial_location)
}
\arguments{
\item{parameters}{A list containing random_seed, random_stream,
human_count, patch_count, and a flow_probability vector.}

\item{initial_location}{A vector saying in which patch each human is located.}
}
\description{
This function returns a flux movement module.
}
\examples{
\dontrun{
movement <- movement_init(params)
time_step <- 0.1
moves <- movement_step(movement, time_step)
person <- human_model(moves, time_step)
}

}
