% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flux_movement_step}
\alias{flux_movement_step}
\title{Time step for a flux movement module.}
\usage{
flux_movement_step(module, time_step)
}
\arguments{
\item{module}{The flux movement list}

\item{time_step}{A floating point time.}
}
\value{
A result list that is opaque.
}
\description{
Time step for a flux movement module.
}
