% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forced_si.R
\name{forced_si_create_bites}
\alias{forced_si_create_bites}
\title{Generate bites for a population of a forced_si model.}
\usage{
forced_si_create_bites(people_cnt, bite_rate, current_time, duration)
}
\arguments{
\item{people_cnt}{How many people there are}

\item{bite_rate}{The number of bites per day.}

\item{current_time}{The start time for bite events}

\item{duration}{in days}
}
\value{
a list of arrays of bite times
}
\description{
Generate bites for a population of a forced_si model.
}
