% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forced_si.R
\name{forced_si_infect}
\alias{forced_si_infect}
\title{Infect an individual from a given set of mosquito bite times.}
\usage{
forced_si_infect()
}
\value{
A list with an enabling function, firing time, and firing function.
}
\description{
We assume the individual state includes a list of times for
infectious bites. This reads that list in order to determine
when they next get bitten.
}
