% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forced_si.R
\name{forced_si_module}
\alias{forced_si_module}
\title{Creates a MASH module that runs forced SI.}
\usage{
forced_si_module(parameters)
}
\arguments{
\item{parameters}{These are simulation parameters. They are
\code{recovery_rate} for recovery, \code{people_cnt} for
the number of individuals, \code{duration_days} for the number
of days per time step. If there are too few or too many parameters,
then this rejects the argument. We disallow extra parameters so that
misnamed parameters don't go unnoticed.}
}
\value{
A simulation object, which is a list.
}
\description{
This module drives infection by supplying bites.
}
