% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forced_si.R
\name{forced_si_observer}
\alias{forced_si_observer}
\title{Observe each transition.}
\usage{
forced_si_observer(transition_name, former_state, new_state, curtime)
}
\arguments{
\item{transition_name}{The string name of the transition.}

\item{former_state}{a list describing the individual's state before the transition}

\item{new_state}{a list describing the individual's state after the transition}

\item{curtime}{The time at which this transition fires.}
}
\value{
a list with the name and time.
}
\description{
The memory use of this method could be improved. We could construct a
data.table into which to store the trajectory, so that it overwrites
lines in the data.table. By returning lists, we're churning memory.
}
