% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forced_si.R
\name{forced_si_population}
\alias{forced_si_population}
\title{Create a population of individuals for forced-SI}
\usage{
forced_si_population(people_cnt, pfpr)
}
\arguments{
\item{people_cnt}{How many people}

\item{pfpr}{What percentage of them should start infected.}
}
\value{
a data.table with columns disease=factor(S, I) and bites,
    which is an empty list.
}
\description{
Create a population of individuals for forced-SI
}
