% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forced_si.R
\name{forced_si_recover}
\alias{forced_si_recover}
\title{Recover an individual from infection}
\usage{
forced_si_recover(rate)
}
\arguments{
\item{rate}{The rate parameter for an exponential distribution}
}
\value{
A list with an enabling function, firing time, and firing function.
}
\description{
Recover an individual from infection
}
