% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_linear.R
\name{health_as_events}
\alias{health_as_events}
\title{Converts wide data table of health events into a long one.}
\usage{
health_as_events(health_dt)
}
\arguments{
\item{health_dt}{One row per individual, two columns per change.
The two columns are for time and state.}

\item{Output}{data table is ID, Level, Time, with a row for each
change for each person, plus starting rows for starting states.}
}
\description{
Converts wide data table of health events into a long one.
}
