% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_constant.R
\name{human_constant_module}
\alias{human_constant_module}
\title{Create human_constant module.}
\usage{
human_constant_module(parameters)
}
\arguments{
\item{parameters}{A list with "recovery_rate", "people_cnt",
"duration_days", "initial_pfpr".}
}
\description{
Create human_constant module.
}
