% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_constant.R
\name{human_disease_path.human_constant}
\alias{human_disease_path.human_constant}
\title{Get events from the human_si module.}
\usage{
\method{human_disease_path}{human_constant}(simulation)
}
\arguments{
\item{simulation}{The module.}
}
\description{
Get events from the human_si module.
}
