% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_loop.R
\name{human_disease_path}
\alias{human_disease_path}
\title{Extract the health status of every person from a health module.}
\usage{
human_disease_path(human_module)
}
\arguments{
\item{human_module}{The module for human health.}
}
\description{
Extract the health status of every person from a health module.
}
