% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\name{human_dwell}
\alias{human_dwell}
\title{Given all human movement, return a matrix of fraction of time at each
location on each day.}
\usage{
human_dwell(movement_dt, day_start, params)
}
\arguments{
\item{movement_dt}{The movement data table where each row is an event.}

\item{params}{A list or data frame that has simulation parameters such
as location_cnt and duration of the time step.}
}
\value{
A three-dimensional array, location x human x days, so that
we can process a day at a time.
}
\description{
Given all human movement, return a matrix of fraction of time at each
location on each day.
}
