% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human-PfSI.R, R/Mosquito-RM.R
\name{human_pfsi_conpars}
\alias{human_pfsi_conpars}
\title{Human-PfSI: Individual Constructor Parameters}
\usage{
human_pfsi_conpars(N, lambda, psi, EIP, p, f, Q, v, M, Y, Z)

human_pfsi_conpars(N, lambda, psi, EIP, p, f, Q, v, M, Y, Z)
}
\arguments{
\item{N}{integer number of patches}

\item{lambda}{numeric matrix (365 x N) describing emergence rate by day of year and patch}

\item{psi}{diffusion (movement) matrix (N x N); rows must sum to 1}

\item{EIP}{integer vector (N) describing length of EIP by day of year}

\item{p}{daily survival probability}

\item{f}{blood feeding rate}

\item{Q}{human blood index}

\item{v}{daily egg production}

\item{M}{initial conditions of total female mosquito populations (N patches)}

\item{Y}{initial conditions of incubating female mosquito populations (N patches)}

\item{Z}{initial conditions of infectious female mosquito populations (N patches)}

\item{id}{integer id of this person}

\item{home_patch_id}{the id of their home patch}

\item{trip_duration}{numeric value; average durarion of trips away from home}

\item{trip_frequency}{numeric value; average rate at which they make trips away from home}

\item{bweight}{numeric value; biting weight to this person (relative to others in a patch; should have mean 1)}

\item{age}{numeric age (not currently used)}

\item{inf}{logical (start simulation infected or not)}

\item{chx}{logical (start simulation with chemoprophylactic protection or not)}
}
\description{
Generates a single named list with parameters to construct a single
'human_pfsi' object (see inst/include/Human-PfSI.hpp for constructor parameters)

Generates a single named list with parameters to construct a
'mosquito_rm' object (see inst/include/Mosquito-RM.hpp for constructor parameters).
Please note that 'N' (number of patches) may not be the same as the total number of patches
constructed in the tile object, due to the existence of reservoir patches. However the
first 1,...,N patches should be non-reservoir patches (simulated explicitly).
}
