% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_si.R
\name{human_si_module}
\alias{human_si_module}
\title{Create human_si module.}
\usage{
human_si_module(parameters)
}
\arguments{
\item{parameters}{A list with "recovery_rate", "people_cnt",
"duration_days", "initial_pfpr".}
}
\description{
Create human_si module.
}
