% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_constant.R
\name{infects_human_path.bloodmeal_constant}
\alias{infects_human_path.bloodmeal_constant}
\title{Extract human bites from the bloodmeal module.}
\usage{
\method{infects_human_path}{bloodmeal_constant}(simulation)
}
\arguments{
\item{simulation}{The bloodmeal_density module.}
}
\value{
A data.table with bite information. These are all bites
of humans where the mosquito was infectious.
}
\description{
Extract human bites from the bloodmeal module.
}
