% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_bites.R
\name{infects_human_path.exponential_bite}
\alias{infects_human_path.exponential_bite}
\title{Get the trajectory from a simulation time step.}
\usage{
\method{infects_human_path}{exponential_bite}(simulation)
}
\arguments{
\item{simulation}{an exponential bites simulation object.}
}
\value{
bites a list with an entry for each person containing a vector of bite times.
}
\description{
Get the trajectory from a simulation time step.
}
