% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\name{infects_mosquito_path.bloodmeal_density}
\alias{infects_mosquito_path.bloodmeal_density}
\title{Extract mosquito bites from the bloodmeal module.}
\usage{
\method{infects_mosquito_path}{bloodmeal_density}(simulation)
}
\arguments{
\item{simulation}{The bloodmeal_linear module.}
}
\value{
A data.table with bite information.
These are only bites where the mosquito was not infectious
but the human was.
}
\description{
Extract mosquito bites from the bloodmeal module.
}
