% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_loop.R
\name{infects_mosquito_path}
\alias{infects_mosquito_path}
\title{Extracts a time series of all bites that can infect a person.}
\usage{
infects_mosquito_path(bloodmeal_module)
}
\description{
An S3 method to read a bloodmeal module and return infectious
bites it has predicted.
}
\seealso{
\code{\link{infects_mosquito_path}}
}
