% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{init_continuous}
\alias{init_continuous}
\title{Given a simulation object, calculate enabled transitions.}
\usage{
init_continuous(simulation)
}
\arguments{
\item{simulation}{This object is created by \code{\link{continuous_simulation}}}
}
\description{
Given a simulation object, calculate enabled transitions.
}
