% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{initialize_times}
\alias{initialize_times}
\title{Set up firing times given an initial state.}
\usage{
initialize_times(individuals, is_enabled, when)
}
\arguments{
\item{individuals}{a data frame of individuals, including columns for times.}

\item{is_enabled}{a list of functions that say when transitions are enabled}

\item{when}{a list of functions that return times when enabled}
}
\value{
a data frame of individuals, with times filled in
}
\description{
Set up firing times given an initial state.
}
