% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{is_enabled}
\alias{is_enabled}
\title{Check to see if transition is enabled for this individual}
\usage{
is_enabled(transition, individual, when, variables = NULL)
}
\arguments{
\item{transition}{A list of transitions}

\item{individual}{A list with individual properties}

\item{when}{A time}

\item{variables}{Any global simulation state}
}
\value{
logical whether transition can fire given this state
}
\description{
This function is for testing the transitions on data for a
single individual. It's a helper function. The running simulation
doesn't call this function.
}
\seealso{
\code{\link{when}}, \code{\link{fire}}
}
