% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{item_dat}
\alias{item_dat}
\title{Item Data}
\format{A data frame with 399 rows and three columns:
\describe{
\item{item_code}{BLS code for the item}
\item{item_name}
\item{display_level}{Level in categorical hierarchy}
}}
\source{
\url{https://download.bls.gov/pub/time.series/cu/cu.item}
}
\usage{
item_dat
}
\description{
Item code, name, and display level
}
\keyword{datasets}
