% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_mosquito.R
\name{lag_mosquito_module}
\alias{lag_mosquito_module}
\title{Constant mosquito count with infections from past bites.}
\usage{
lag_mosquito_module(parameters)
}
\arguments{
\item{parameters}{containing \code{people_cnt}, \code{bite_rate}, \code{duration}.}
}
\value{
A bite module.
}
\description{
This assumes the number of mosquitos in a place is constant.
The number of infectious mosquitoes is the number of incoming infectious bites
from \code{lag_days} ago, times the \code{decline_rate} parameter.
}
\details{
On initialization, it uses a given start number, \code{initial_prevalence}, until the simulation has
progressed \code{lag_days}.
}
