% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{local_logging}
\alias{local_logging}
\title{Set up logging for use on a local machine.}
\usage{
local_logging(level_name = "info", logfn = NULL)
}
\arguments{
\item{level_name}{A string name from \code{log_levels()}.}

\item{logfn}{A logfile name, default is NULL which means no log file.}
}
\description{
Set up logging for use on a local machine.
}
