% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_loop.R
\name{location_path}
\alias{location_path}
\title{Extracts a time series, by location, of which people are present.}
\usage{
location_path(location_module)
}
\description{
An S3 method to read result data from a location module, where
that data is indexed first by location and then by people present over time.
}
\seealso{
\code{\link{person_path}}
}
