% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletrip.R
\name{location_path.simple_trip}
\alias{location_path.simple_trip}
\title{Return trajectory by location for Simple Trip model}
\usage{
\method{location_path}{simple_trip}(simulation)
}
\arguments{
\item{simulation}{A simple trip model (most likely made via \code{\link[macro]{simple_trip_module}})}
}
\value{
a list of \code{data.frame}, one for each patch giving the times anyone arrived at or left that patch
}
\description{
Return trajectory by location for Simple Trip model
}
