% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_location.R
\name{location_path.single_location}
\alias{location_path.single_location}
\title{Get the data for where everybody has been. Surprise, it's one place.}
\usage{
\method{location_path}{single_location}(module)
}
\arguments{
\item{module}{The single_location module from which to extract results.}
}
\value{
a list of who was at the one place, everybody.
}
\description{
Get the data for where everybody has been. Surprise, it's one place.
}
