% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{look_back_eip}
\alias{look_back_eip}
\title{This reverses the list of EIP per day of the year.}
\usage{
look_back_eip(EIP)
}
\arguments{
\item{EIP}{a vector of the EIP on a day of the year.

It tells you how many days back to get the newly-infected
mosquitoes. The value is a ragged array by day of year.}
}
\description{
This reverses the list of EIP per day of the year.
}
