% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{major_groups_dat}
\alias{major_groups_dat}
\title{Major Group Data (and meta data)}
\format{An object of class \code{data.frame} with 6517 rows and 6 columns.}
\usage{
major_groups_dat
}
\description{
Major Group Data (and meta data)
}
\keyword{datasets}
