% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{make_tosolve}
\alias{make_tosolve}
\title{Find fraction of mosquitoes bitten each day, given fraction infected.}
\usage{
make_tosolve(parameters, location_idx)
}
\arguments{
\item{parameters}{The input module parameters.}
}
\value{
A function to use for optimization.
}
\description{
Find fraction of mosquitoes bitten each day, given fraction infected.
}
