% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_bites.R
\name{mash_step.exponential_bite}
\alias{mash_step.exponential_bite}
\title{One time step of an exponential bite module.}
\usage{
\method{mash_step}{exponential_bite}(simulation)
}
\arguments{
\item{simulation}{the simulation object.}
}
\value{
a modified simulation object
}
\description{
One time step of an exponential bite module.
}
