% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forced_si.R
\name{mash_step.forced_si}
\alias{mash_step.forced_si}
\title{Takes one time step of the discrete time step.}
\usage{
\method{mash_step}{forced_si}(simulation, step_id, bites)
}
\arguments{
\item{simulation}{A forced-SI model.}

\item{bites}{A datatable where each row has a human=ID
and a bites=list(array of bite times).}
}
\value{
a simulation object
}
\description{
Takes one time step of the discrete time step.
}
\examples{
\dontrun{
step_si_module(simulation,
               forced_si_create_bites(100, 1/20, current_time, 14))
}
}
