% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_constant.R
\name{mash_step.human_constant}
\alias{mash_step.human_constant}
\title{One step of the human_constant module.}
\usage{
\method{mash_step}{human_constant}(simulation, step_id, bites_dt)
}
\arguments{
\item{simulation}{The module.}

\item{bites_dt}{a data.table of bites.
This ignores its imput and updates the time.}
}
\description{
One step of the human_constant module.
}
