% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_si.R
\name{mash_step.human_si}
\alias{mash_step.human_si}
\title{One step of the human_si module.}
\usage{
\method{mash_step}{human_si}(simulation, step_id, bites_dt)
}
\arguments{
\item{simulation}{The module.}

\item{bites_dt}{a data.table of bites.}
}
\description{
One step of the human_si module.
}
