% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_mosquito.R
\name{mash_step.lag_mosquito}
\alias{mash_step.lag_mosquito}
\title{One time step of an exponential bite module.}
\usage{
\method{mash_step}{lag_mosquito}(simulation, step_id, bloodmeal_path)
}
\arguments{
\item{simulation}{the simulation object.}

\item{bloodmeal_path}{A list of bites that are infectious to mosquitoes and times.
These are the bites received during the last time step, not this one.}
}
\value{
a modified simulation object
}
\description{
One time step of an exponential bite module.
}
