% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mash_step.mosquito_rm}
\alias{mash_step.mosquito_rm}
\title{One discrete time step for the mosquito_rm module.}
\usage{
\method{mash_step}{mosquito_rm}(module, bloodmeal_dt)
}
\arguments{
\item{module}{The mosquito_rm module.}

\item{bloodmeal_dt}{The input bloodmeal data from the last time step.
This has Bites, Time, Location in a data table.}
}
\value{
The mosqutio_rm module back again, after the time step.
}
\description{
One discrete time step for the mosquito_rm module.
}
