% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_loop.R
\name{mash_step}
\alias{mash_step}
\title{Increments a MASH module by one discrete time step.}
\usage{
mash_step(mash_module, ...)
}
\description{
An S3 method to advance a module by whatever is the simulation
time step.
}
