% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_location.R
\name{mash_step.single_location}
\alias{mash_step.single_location}
\title{Increment time so that everybody stays in the one location.}
\usage{
\method{mash_step}{single_location}(location_module, health_path)
}
\arguments{
\item{location_module}{The MASH module.}

\item{health_path}{The history of health states, which is ignored here.}
}
\description{
Increment time so that everybody stays in the one location.
}
