% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symtable.R
\name{mclear}
\alias{mclear}
\title{Macro Clear Function}
\usage{
mclear()
}
\value{
The number of objects cleared, invisibly.  The function also
outputs a message saying how many objects were cleared.
}
\description{
The \code{mclear} function clears the macro symbol table of any
stored macro variables and macro functions.  The function is used
to avoid contamination between one call to \code{msource} and the
next.  It is called automatically when the "clear" parameter of
\code{msource} is set to TRUE.  The function takes no parameters.
}
