% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_dat}
\alias{meta_dat}
\title{Meta Data}
\format{6745 x 13 data frame}
\source{
\url{https://download.bls.gov/pub/time.series/cu/cu.series}
}
\usage{
meta_dat
}
\description{
Meta Data
}
\keyword{datasets}
