% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msource.R
\name{mget}
\alias{mget}
\title{Macro Get Function}
\usage{
mget(name)
}
\arguments{
\item{name}{The name of the macro variable as a quoted string, with no
leading ampersand or trailing dot (".").  The leading ampersand will
be added automatically by the function.  This
parameter is required.}
}
\value{
The value of the macro variable as a character string.  If the variable
name is not found, the function will return an NA.
}
\description{
The \code{mget} function extracts the value of a single macro variable
from the macro symbol table.
}
\examples{
library(macro)

# Get path to demo macro program
src <- system.file("extdata/Demo3.R", package = "macro")

# Create temporary output path
tmp <- file.path(tempdir(), "Demo3_mod.R")

# Display source code
# - This is the macro input code
cd <- readLines(src)
cat(paste(cd, "\n"))
# #\% Determine appropriate data path
# #\%if ("&env." == "prod")
#   #\%let pth <- /projects/prod/data
# #\%else
#   #\%let pth <- /projects/dev/data
# #\%end

# Set environment variable using mset()
mset("env", "prod")

# Macro Execute Source Code
# - set clear to FALSE to so "env" value is not removed
msource(src, tmp, clear = FALSE)

# View "pth" macro variable
res <- mget("pth")

# View results
res
# [1] "/projects/prod/data"

}
