% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_mosquito.R
\name{mosquito_path.lag_mosquito}
\alias{mosquito_path.lag_mosquito}
\title{Get the trajectory from a simulation time step.}
\usage{
\method{mosquito_path}{lag_mosquito}(simulation)
}
\arguments{
\item{simulation}{an exponential bites simulation object.}
}
\value{
bites a list with an entry for each person containing a vector of bite times.
}
\description{
Get the trajectory from a simulation time step.
}
