% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mosquito_path.mosquito_rm}
\alias{mosquito_path.mosquito_rm}
\title{Get the output of the time step for Ross Macdonald mosquitoes.}
\usage{
\method{mosquito_path}{mosquito_rm}(module)
}
\arguments{
\item{module}{A \code{mosquito_rm} module.}
}
\value{
A data.table
}
\description{
Get the output of the time step for Ross Macdonald mosquitoes.
}
