% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mosquito_rm_aquatic}
\alias{mosquito_rm_aquatic}
\title{An aquatic emergence dynamics.}
\usage{
mosquito_rm_aquatic(lambda)
}
\arguments{
\item{A}{vector of emergence rate per patch.}
}
\value{
Returns a vector count of emerged mosquitoes.
}
\description{
An aquatic emergence dynamics.
}
