% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mosquito_rm_average_kappa}
\alias{mosquito_rm_average_kappa}
\title{An average of kappa over the past ten days so that we can project forward.}
\usage{
mosquito_rm_average_kappa(kappa)
}
\arguments{
\item{kappa}{A matrix of values for each patch, for all days. Nx10.}
}
\value{
A matrix of kappa values for each patch and day, Nx10.

We use this function because it can look at trends instead of a
simple average. This version weights most recent days to construct
a constant value for the future. These values shouldn't have an
effect on the module's output because of delays in the system,
but we make them to be consistent.
}
\description{
An average of kappa over the past ten days so that we can project forward.
}
