% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosquito-RM.R
\name{mosquito_rm_conpars}
\alias{mosquito_rm_conpars}
\title{Mosquito-RM: Constructor Parameters}
\usage{
mosquito_rm_conpars(N, lambda, psi, EIP, p = 0.9, f = 0.3, Q = 0.9,
  v = 20, M, Y, Z)
}
\arguments{
\item{N}{integer number of patches}

\item{lambda}{numeric matrix (365 x N) describing emergence rate by day of year and patch}

\item{psi}{diffusion (movement) matrix (N x N); rows must sum to 1}

\item{EIP}{integer vector (N) describing length of EIP by day of year}

\item{p}{daily survival probability}

\item{f}{blood feeding rate}

\item{Q}{human blood index}

\item{v}{daily egg production}

\item{M}{initial conditions of total female mosquito populations (N patches)}

\item{Y}{initial conditions of incubating female mosquito populations (N patches)}

\item{Z}{initial conditions of infectious female mosquito populations (N patches)}
}
\description{
Generates a single named list with parameters to construct a
'mosquito_rm' object (see inst/include/Mosquito-RM.hpp for constructor parameters).
Please note that 'N' (number of patches) may not be the same as the total number of patches
constructed in the tile object, due to the existence of reservoir patches. However the
first 1,...,N patches should be non-reservoir patches (simulated explicitly).
This model needs a log stream with name 'mosquito'.
}
