% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mosquito_rm_copy_state}
\alias{mosquito_rm_copy_state}
\title{Make a copy of the state of the Mosquito-RM model.}
\usage{
mosquito_rm_copy_state(state)
}
\arguments{
\item{state}{A list of the state}
}
\value{
A copy of the state with no changes.
}
\description{
Make a copy of the state of the Mosquito-RM model.
}
