% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mosquito_rm_discrete_step}
\alias{mosquito_rm_discrete_step}
\title{One discrete time step for the mosquito_rm module.}
\usage{
mosquito_rm_discrete_step(module, past_kappa)
}
\arguments{
\item{module}{The mosquito_rm module.}

\item{past_kappa}{Infectiousness to mosquitoes over discrete-time step.}
}
\value{
Observation of the time step and new state as a list
with \code{future_state} and \code{output}.

This module receives correct input for one ten-day time step
and returns correct output for the next ten-day time step.
To do that, it saves its state after the first ten-day calculation
and recalculates the second set of ten days the next time
it is called.
}
\description{
One discrete time step for the mosquito_rm module.
}
