% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mosquito_rm_module}
\alias{mosquito_rm_module}
\title{Create a Mosquito_rm module for Ross-Macdonald mosquitoes}
\usage{
mosquito_rm_module(parameters)
}
\arguments{
\item{parameters}{A block of parameters for this model.}
}
\value{
A new module.
}
\description{
Create a Mosquito_rm module for Ross-Macdonald mosquitoes
}
