% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mosquito_rm_trended_kappa}
\alias{mosquito_rm_trended_kappa}
\title{A trend of kappa over the past ten days so that we can project forward.}
\usage{
mosquito_rm_trended_kappa(kappa)
}
\arguments{
\item{kappa}{A matrix of values for each patch, for all days. Nx10.}
}
\value{
A matrix of kappa values for each patch and day, Nx10.

We use this function because it can look at trends instead of a
simple average. This version uses an autoregressive process to
find a trend.
}
\description{
A trend of kappa over the past ten days so that we can project forward.
}
