% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{movements_of_human}
\alias{movements_of_human}
\title{Extract human movements from the result of a movement step.}
\usage{
movements_of_human(movement_list, human)
}
\arguments{
\item{movement_list}{The movement result object.}

\item{human}{A list of the humans for whom you want the movements.}
}
\value{
A vector that's two wide.
}
\description{
Extract human movements from the result of a movement step.
}
