% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msource.R
\name{mput}
\alias{mput}
\title{Macro Put Function}
\usage{
mput()
}
\value{
A data frame of macro symbols and their values.
}
\description{
The \code{mput} function extracts the contents of the macro symbol table
as a data frame, and, by default, displays it in the console.
The function can be used to examine the symbol table values.
}
