% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mrm_copy_state}
\alias{mrm_copy_state}
\title{Make a copy of the state of the Mosqutio-RM model.}
\usage{
mrm_copy_state(state)
}
\arguments{
\item{state}{A list of the state}
}
\value{
A copy of the state. No changes.
}
\description{
Make a copy of the state of the Mosqutio-RM model.
}
