% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msource.R
\name{mset}
\alias{mset}
\title{Macro Set Function}
\usage{
mset(x, value = NULL)
}
\arguments{
\item{x}{The name of the macro variable to assign,
passed as a quoted string with no leading ampersand or trailing
dot (".").  The leading ampersand will be added automatically
by the function.  This parameter is required.}

\item{value}{The value of the macro variable to assign.
Value will be converted to a character string. This parameter is
not required.  If the value parameter is not supplied, the variable
will be removed from the symbol table.}
}
\value{
The macro name, invisibly.
}
\description{
The \code{mset} function assigns the value of a macro
variable from regular R code.
}
\examples{
library(macro)

# Get path to demo macro program
src <- system.file("extdata/Demo3.R", package = "macro")

# Create temporary output path
tmp <- file.path(tempdir(), "Demo3_mod.R")

# Display source code
# - This is the macro input code
cd <- readLines(src)
cat(paste(cd, "\n"))
# #\% Determine appropriate data path
# #\%if ("&env." == "prod")
#   #\%let pth <- /projects/prod/data
# #\%else
#   #\%let pth <- /projects/dev/data
# #\%end

# Set environment variable using mset()
mset("env", "prod")

# Macro Execute Source Code
# - set clear to FALSE to so "env" value is not removed
msource(src, tmp, clear = FALSE)

# View "pth" macro variable
res <- mget("pth")

# View results
res
# [1] "/projects/prod/data"

}
