% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_equations.R
\name{normalize_equations}
\alias{normalize_equations}
\title{This takes a list of equations and sorts them into a standard form.}
\usage{
normalize_equations(equations)
}
\arguments{
\item{equations}{This is a list of what R calls formulas.

This assumes the equations are linear with coefficients that can
be any expression. The left-hand side of the equation, where the
derivative would go, should just be the variable itself.}
}
\description{
This takes a list of equations and sorts them into a standard form.
}
\examples{
normalize_equations(list(
  U ~ -h * U + r * X,
  X ~ h * U - r * X
))
}
