% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_begin_step.nothing_observer}
\alias{observe_begin_step.nothing_observer}
\title{Called at beginning of the time step for any saving of observations.}
\usage{
\method{observe_begin_step}{nothing_observer}(observer, step)
}
\description{
Called at beginning of the time step for any saving of observations.
}
