% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_begin_step}
\alias{observe_begin_step}
\title{Called at beginning of the time step for any saving of observations.}
\usage{
observe_begin_step(observer, step)
}
\description{
Called at beginning of the time step for any saving of observations.
}
