% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_bloodmeal_human.complete_observer}
\alias{observe_bloodmeal_human.complete_observer}
\title{Observe the bloodmeal bites to humans.}
\usage{
\method{observe_bloodmeal_human}{complete_observer}(observer, bloodmeal_human, step)
}
\description{
Observe the bloodmeal bites to humans.
}
