% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_bloodmeal_human}
\alias{observe_bloodmeal_human}
\title{Observe the bloodmeal bites to humans.}
\usage{
observe_bloodmeal_human(observer, bloodmeal_path, step)
}
\description{
Observe the bloodmeal bites to humans.
}
