% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_bloodmeal_mosquito.complete_observer}
\alias{observe_bloodmeal_mosquito.complete_observer}
\title{Observe the bloodmeal bites to mosquitoes.}
\usage{
\method{observe_bloodmeal_mosquito}{complete_observer}(observer, bloodmeal_mosquito, step)
}
\description{
Observe the bloodmeal bites to mosquitoes.
}
