% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{observe_bloodmeal_mosquito.mosquito_rm}
\alias{observe_bloodmeal_mosquito.mosquito_rm}
\title{Get the output of the time step for Ross Macdonald mosquitoes.}
\usage{
observe_bloodmeal_mosquito.mosquito_rm(module)
}
\arguments{
\item{module}{A \code{mosquito_rm} module.}
}
\value{
A data.table with three columns, \code{Bites}, \code{Location}, and
\code{Time}.
}
\description{
Get the output of the time step for Ross Macdonald mosquitoes.
}
