% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_loop.R, R/observer.R
\name{observe_bloodmeal_mosquito}
\alias{observe_bloodmeal_mosquito}
\title{Observe data from mosquitoes that goes to the bloodmeal.}
\usage{
observe_bloodmeal_mosquito(observer, bloodmeal_path, step)

observe_bloodmeal_mosquito(observer, bloodmeal_path, step)
}
\arguments{
\item{mosquito_module}{The module for mosquito life cycle.}
}
\description{
Observe data from mosquitoes that goes to the bloodmeal.

Observe the bloodmeal bites to mosquitoes.
}
