% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_end_step.nothing_observer}
\alias{observe_end_step.nothing_observer}
\title{Called at end of the time step for any saving of observations.}
\usage{
\method{observe_end_step}{nothing_observer}(observer, step)
}
\description{
Called at end of the time step for any saving of observations.
}
