% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_end_step}
\alias{observe_end_step}
\title{Called at end of the time step for any saving of observations.}
\usage{
observe_end_step(observer, step)
}
\description{
Called at end of the time step for any saving of observations.
}
