% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_health.complete_observer}
\alias{observe_health.complete_observer}
\title{Observe the health state for every person.}
\usage{
\method{observe_health}{complete_observer}(observer, health_path, step)
}
\description{
Observe the health state for every person.
}
