% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_location.nothing_observer}
\alias{observe_location.nothing_observer}
\title{Observe the change in location for every person.}
\usage{
\method{observe_location}{nothing_observer}(observer, location_path, step)
}
\description{
Observe the change in location for every person.
}
