% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_location}
\alias{observe_location}
\title{Observe the change in location for every person.}
\usage{
observe_location(observer, location_path, step)
}
\description{
Observe the change in location for every person.
}
