% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_mosquito.nothing_observer}
\alias{observe_mosquito.nothing_observer}
\title{Observe the mosquito state.}
\usage{
\method{observe_mosquito}{nothing_observer}(observer, mosquito_path, step)
}
\description{
Observe the mosquito state.
}
