% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACRO-Utility.R
\name{pars_obj}
\alias{pars_obj}
\title{Utility: Make a Parameter Object for C++ Initialization}
\usage{
pars_obj(val, name)
}
\arguments{
\item{val}{numeric or integer value}

\item{name}{name of the parameter}
}
\description{
Take a parameter and transform into a package (named list) so it can
easily be passed to the parameters object in C++ (see inst/include/Parameters.hpp)
}
