% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch.R
\name{patch_conpars}
\alias{patch_conpars}
\title{Patch: Constructor Parameters}
\usage{
patch_conpars(id, move, bWeightZoo, bWeightZootox, reservoir, res_EIR)
}
\arguments{
\item{id}{integer id}

\item{move}{'row' of the stochastic movement matri}

\item{bWeightZoo}{biting weight assigned to non-human hosts}

\item{bWeightZootox}{biting weight assigned to endectocide-treated non-human hosts}

\item{reservoir}{boolean; is this patch a reservoir? (no simulation of mosquitos)}

\item{res_EIR}{fixed EIR for reservoir patch}
}
\description{
Generates a single named list with parameters to construct a
'patch' object (see inst/include/Patch.hpp for constructor parameters).
}
