% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch.R
\name{patches_parameters}
\alias{patches_parameters}
\title{Patch: Constructor Parameters for all Patches}
\usage{
patches_parameters(move, bWeightZoo, bWeightZootox, reservoir, res_EIR)
}
\arguments{
\item{move}{stochastic movement matrix}

\item{bWeightZoo}{biting weight assigned to non-human hosts}

\item{bWeightZootox}{biting weight assigned to endectocide-treated non-human hosts}

\item{reservoir}{boolean; is this patch a reservoir? (no simulation of mosquitos)}

\item{res_EIR}{fixed EIR for reservoir patch}
}
\description{
Generates a list with parameters to construct all patches by calling \code{\link{patch_conpars}}
repeatedly. Note that vectors 'reservoir' and 'res_EIR' must be given for all patches (even non-reservoir patches)
}
