% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_flux.R
\name{person_path.flux}
\alias{person_path.flux}
\title{Return trajectory by person for Flux model}
\usage{
\method{person_path}{flux}(simulation)
}
\arguments{
\item{simulation}{A flux model (most likely made via \code{\link[macro]{flux_module}})}
}
\value{
a list of \code{data.frame}, one for person giving the time and new location when they moved
}
\description{
Return trajectory by person for Flux model
}
