% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_loop.R
\name{person_path}
\alias{person_path}
\title{Extracts a time series, for each person, of their locations over time.}
\usage{
person_path(location_module)
}
\description{
An S3 method to read result data from a location module,
where that data is indexed by person, making it easy to look up
where that person has been over time.
}
\seealso{
\code{\link{location_path}}
}
