% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_location.R
\name{person_path.single_location}
\alias{person_path.single_location}
\title{Retrieve list of where people have gone.}
\usage{
\method{person_path}{single_location}(simulation)
}
\arguments{
\item{simulaton}{A module of class single_location.}
}
\description{
Retrieve list of where people have gone.
}
