% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PfSI-Utility.R
\name{pfsi_human_output}
\alias{pfsi_human_output}
\title{PfSI: Process Human Infection Output}
\usage{
pfsi_human_output(h_inf)
}
\arguments{
\item{h_inf}{a \code{data.frame} read in by \code{read.csv(file = path,stringsAsFactors = FALSE)}}
}
\description{
After reading in the .csv output for the \code{human_inf} logging, bin humans
by states with a resolution of one day for analysis and plotting.
This function ignores information on location.
}
