% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_dat}
\alias{reg_dat}
\title{Top 25, for lm(...)}
\format{An object of class \code{data.frame} with 1263 rows and 27 columns.}
\usage{
reg_dat
}
\description{
The top 25 item strata most correlated with CPI,
formatted for fitting regression models
}
\keyword{datasets}
