% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riws}
\alias{riws}
\title{CPI Relative Importance Weights}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 322 rows and 3 columns.}
\usage{
riws
}
\description{
CPI Relative Importance Weights
}
\keyword{datasets}
