% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riws2009}
\alias{riws2009}
\title{2009 Strata RIWs}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 68 rows and 4 columns.}
\source{
\url{https://www.bls.gov/cpi/tables/relative-importance/home.htm}

\url{https://www.bls.gov/cpi/tables/relative-importance/2009.txt}
}
\usage{
riws2009
}
\description{
The 2009 Relative Importance Weights for all 68 strata.
}
\keyword{datasets}
