% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{run_continuous}
\alias{run_continuous}
\title{Run the simulation for a certain time.}
\usage{
run_continuous(simulation, duration)
}
\arguments{
\item{simulation}{This object is created by
\code{\link{continuous_simulation}}}

\item{duration}{How long to run}
}
\value{
The trajectory of the run.
}
\description{
Run the simulation for a certain time.
}
