% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_macro}
\alias{run_macro}
\title{Run a MACRO Simulation}
\usage{
run_macro(seed, tmax, human_pars, mosquito_pars, patch_pars, model_pars,
  log_streams, vaxx_events, verbose = TRUE)
}
\arguments{
\item{seed}{a non-negative integer seed for the prng (the 'mt19937' implementation of the Mersenne Twister is used; see \url{http://www.cplusplus.com/reference/random/mt19937/})}

\item{tmax}{a non-negative integer maximum simulation time (simulation runs from day 0 to tmax-1)}

\item{human_pars}{a list of constructor parameters for each human in the simulation}

\item{mosquito_pars}{a list of constructor parameters for the mosquito population}

\item{patch_pars}{a list of constructor parameters for each patch in the simulation}

\item{model_pars}{a list of parameter objects (see \code{\link{pars_obj}}) to initialize parameters object (see inst/include/Parameters.hpp for details)}

\item{log_streams}{a list giving the output file and name of each logging stream (please ensure that you have provided the right logging stream names for the model(s) you wish to run)}

\item{vaxx_events}{a list of vaccination events on the human population (for no vaccination events, pass an empty list, \code{list()})}

\item{verbose}{print information to console?}
}
\description{
write me!
}
