% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{runMSource}
\alias{runMSource}
\title{Addin Function to Run msource()}
\usage{
runMSource()
}
\value{
The results of \code{msource}, invisibly.
}
\description{
This function is exposed to the addin menu to run \code{msource()}
interactively. The function will run either the
currently selected code, or the currently active program.
The function sets the "envir" parameter
to the global environment and the "clear" parameter to FALSE
to enhance the user experience.  The function
takes no parameters and is only used by the addin menu.
}
\examples{
# Called from addin menu
# runMSource()
}
