% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{runMSourceDebug}
\alias{runMSourceDebug}
\title{Addin Function to Run msource() in Debug Mode}
\usage{
runMSourceDebug()
}
\description{
This function is exposed to the addin menu to run \code{msource()}
in debug mode.
}
