% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_sample.R
\name{sample_health_infection_status}
\alias{sample_health_infection_status}
\title{Creates a dataset that health would send to bloodmeal.}
\usage{
sample_health_infection_status(human_cnt = 10L, time_step = 10)
}
\arguments{
\item{human_cnt}{The number of humans as an integer.}

\item{time_step}{The length of a time step as a numeric.

You can use this to get a look at what the data would be.}
}
\description{
Creates a dataset that health would send to bloodmeal.
}
