% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_sample.R
\name{sample_humans_at_location}
\alias{sample_humans_at_location}
\title{Sample of human activity at a location, within the bloodmeal.}
\usage{
sample_humans_at_location(human_cnt = 10L, move_period = 20, time_step = 10)
}
\arguments{
\item{human_cnt}{Number of humans.}

\item{move_period}{Not used}

\item{time_step}{Defaults to 10.}
}
\value{
A data.table with the columns \code{ID}, \code{Level}, \code{Time}
\code{Location}, and \code{Event}. The location will be 1 for all
rows. The event is 1 for enter, -1 for leave and 2 for
change in infection status. The level is 0 or 1.
}
\description{
Sample of human activity at a location, within the bloodmeal.
}
