% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_sample.R
\name{sample_mosquito_half_bites}
\alias{sample_mosquito_half_bites}
\title{Create a sample dataset from mosquitoes to bloodmeal.}
\usage{
sample_mosquito_half_bites(bite_cnt = 10L, place_cnt = 3L, time_step = 10)
}
\arguments{
\item{bite_cnt}{Integer number of bites.}

\item{place_cnt}{Integer number of locations to do the biting.}

\item{time_step}{Duration of time step within which to bite.}
}
\description{
Create a sample dataset from mosquitoes to bloodmeal.
}
