% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_sample.R
\name{sample_mosquito_kappa}
\alias{sample_mosquito_kappa}
\title{Create a sample dataset from mosquitoes to bloodmeal.}
\usage{
sample_mosquito_kappa(place_cnt = 3L, time_step = 10)
}
\arguments{
\item{place_cnt}{Integer number of locations to do the biting.}

\item{time_step}{Duration of time step within which to bite.}
}
\description{
Create a sample dataset from mosquitoes to bloodmeal.
}
