% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_sample.R
\name{sample_mosquito_myz}
\alias{sample_mosquito_myz}
\title{Sample dataset output from mosquito to bloodmeal using (M, Y, Z).}
\usage{
sample_mosquito_myz(place_cnt = 3L, time_step = 10)
}
\arguments{
\item{place_cnt}{Integer number of locations to do the biting.}

\item{time_step}{Duration of time step within which to bite.}
}
\description{
Sample dataset output from mosquito to bloodmeal using (M, Y, Z).
}
