% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_daily.R
\name{sample_mosquitoes}
\alias{sample_mosquitoes}
\title{Creates a population of mosquitoes for a given day.}
\usage{
sample_mosquitoes(world)
}
\arguments{
\item{world}{The parameters for the simulate}
}
\description{
Creates a population of mosquitoes for a given day.
}
