% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_sample.R
\name{sample_move_location}
\alias{sample_move_location}
\title{Creates a sample movement dataset, that would be sent to the bloodmeal module.}
\usage{
sample_move_location(human_cnt = 10L, place_cnt = 5L, time_step = 10)
}
\arguments{
\item{human_cnt}{The number of people as an integer.}

\item{place_cnt}{The number of possible locations as an integer.}

\item{time_step}{The length of a time step as a numeric.}
}
\description{
Creates a sample movement dataset, that would be sent to the bloodmeal module.
}
