% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_daily.R
\name{sample_travel}
\alias{sample_travel}
\title{Given a probability to be in each place, sample human movement.}
\usage{
sample_travel(world)
}
\arguments{
\item{world}{Parameters that include a travel pattern matrix.}
}
\value{
A matrix of dwell locations for each person.
}
\description{
Given a probability to be in each place, sample human movement.
}
