% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{set_mersenne_seed}
\alias{set_mersenne_seed}
\title{Set the seed on the Mersenne Twister random generator.}
\usage{
set_mersenne_seed(seed_vector)
}
\arguments{
\item{seed_vector}{An array of seed values.}
}
\description{
The input is a vector of 624 integers, used to initialize
an exact state for MT19937. We assume it comes from
}
\details{
copies <- skip_mersenne_twister(4789, 17, 100)
this_run <- 1 # Out of 100 copies we run.
set_mersenne_seed(copies[, this_run])
}
