% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{shift_with_open_interval}
\alias{shift_with_open_interval}
\title{Makes a shift matrix that accumulates the oldest.}
\usage{
shift_with_open_interval(N)
}
\arguments{
\item{N}{The size of the square matrix.}
}
\value{
A square matrix. You right-multiply a population, and it gets
a day older by moving a column to the right.

This is for half-open intervals, meaning the last column of the matrix
doesn't shift. It accumulates individuals that are added from the previous
day. This number will be gradually reduced by the survival.
}
\description{
Makes a shift matrix that accumulates the oldest.
}
