% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_simpletrip.R
\name{simple_trip_module}
\alias{simple_trip_module}
\title{Creates a MASH module that runs Simple Trip}
\usage{
simple_trip_module(parameters)
}
\arguments{
\item{parameters}{A list or environment}
}
\value{
A simulation object, which is a list.
}
\description{
This module simulates movement according to a 'simple trip' model, where everybody
has a home which they leave to go on trips according to some rate. When they arrive at
a place they stay there for some time before returning home, upon which they schedule the
next trip.
}
\details{
The input \code{parameters} must have the following \emph{named} members:
\itemize{
\item \code{trip_rate}: the rate at which each person takes a trip away from home
\item \code{trip_dest}: the stochastic matrix telling us where each person (row) goes
(col) when they leave home
\item \code{return_home_rate}: the rate at which each person (row) goes home from their current location (col)
\item \code{npatch}: number of patches/places
\item \code{home}: vector of everybody's home
\item \code{current}: vector of everybody's current location
}
}
