% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletrip.R
\name{simple_trip_observer}
\alias{simple_trip_observer}
\title{Movement: Simple 'Simple Trip' Observer}
\usage{
simple_trip_observer(transition_name, former_state, new_state, curtime)
}
\arguments{
\item{transition_name}{The string name of the transition.}

\item{former_state}{a list describing the individual's state before the transition}

\item{new_state}{a list describing the individual's state after the transition}

\item{curtime}{The time at which this transition fires.}
}
\value{
a list with the name and time.
}
\description{
The memory use of this method could be improved. We could construct a
data.table into which to store the trajectory, so that it overwrites
lines in the data.table. By returning lists, we're churning memory.
}
\details{
Currently this is identical to \code{\link[macro]{forced_si_observer}}

You may want to augment this to record the id of the individual.
}
