% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletrip.R
\name{simple_trip_stateoutput}
\alias{simple_trip_stateoutput}
\title{Movement: Simple Trip Trajectory to State occupancy}
\usage{
simple_trip_stateoutput(trajectory, init_state)
}
\arguments{
\item{trajectory}{the output of \code{\link[macro]{run_continuous}}}

\item{init_state}{the initial state (S1,S2,S3,S4)}
}
\value{
a vector of time-averaged state occupancy probabilities
}
\description{
This is an internal function used to verify that the 'simple trip'
model is functioning correctly. See the simple trip vignette
for more details.
}
