% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_simpletrip.R
\name{simple_trip_transitions}
\alias{simple_trip_transitions}
\title{Movement: Make 'Simple Trip' Transitions}
\usage{
simple_trip_transitions()
}
\value{
a list of transitions
}
\description{
This function makes the set of transitions for the 'simple trip' model of movement;
the simple trip model is the simplest Lagrangian (non-Eulerian) movement model, meaning that it
is not a flux type model of movement. Each person has a home, so for \eqn{N}
patches the state space is \eqn{N^{2}} as we track what patches residents of patch \eqn{i}
are at.
}
