% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\name{single_dwell}
\alias{single_dwell}
\title{Given the movement history of a single person, calculate fraction of
time in any place.}
\usage{
single_dwell(
  h_record,
  location_cnt,
  move_cnt,
  day_start,
  step_duration,
  day_duration = 1
)
}
\arguments{
\item{h_record}{A row from the movement table.}

\item{location_cnt}{The number of locations.}

\item{move_cnt}{The maximum number of moves possible (width of movement table/2)}

\item{step_duration}{Amount of time for all time steps.}

\item{day_duration}{How long each time step is.}
}
\value{
an array with fraction of time in each location on each day.
The size is location x days.
}
\description{
Given the movement history of a single person, calculate fraction of
time in any place.
}
