% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_equations.R
\name{single_jump}
\alias{single_jump}
\title{Use equations to make a function that calculates the matrix for a single jump.}
\usage{
single_jump(equations)
}
\arguments{
\item{equations}{A list of formulas.}
}
\description{
Use equations to make a function that calculates the matrix for a single jump.
}
\seealso{
\code{\link{normalize_equations}}
}
