% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_location.R
\name{single_location}
\alias{single_location}
\title{MASH module that says everybody is in one location.}
\usage{
single_location(params)
}
\description{
MASH module that says everybody is in one location.
}
