% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skip_mersenne_twister}
\alias{skip_mersenne_twister}
\title{Generate Mersenne Twister seed states skipped ahead by powers of two.}
\usage{
skip_mersenne_twister(seed, skip_size_power_two, skip_count)
}
\arguments{
\item{seed}{A single number, without too many 0 bits, to
initialize the random number generator.}

\item{skip_size_power_two}{The jump will be two raised to this
power. If it is 14 or greater, it will use the
special algorithm.}

\item{skip_count}{How many times to skip forward. This is
the number of simultaneous streams you want to use.}
}
\value{
A matrix of size 624 by skip_count. Put the 624
        into the Mersenne seed in order to initialize it.
}
\description{
A Mersenne Twister random number generator can be skipped ahead
by large numbers of random numbers. This function uses the Boost
implementation of that algorithm to skip ahead its version of
mt19937, not mt19937_64.
}
