% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_equations.R
\name{small_step_maker}
\alias{small_step_maker}
\title{Use equations to make a function that calculates the matrix for small time steps.}
\usage{
small_step_maker(equations)
}
\arguments{
\item{equations}{A list of formulas.}
}
\description{
Use equations to make a function that calculates the matrix for small time steps.
}
\seealso{
\code{\link{normalize_equations}}
}
