% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move-module.R
\name{step.bites_multinom_machine}
\alias{step.bites_multinom_machine}
\title{Distribute infectious Bites to Humans (Multinomial)}
\usage{
step.bites_multinom_machine(parameters, time_step)
}
\arguments{
\item{parameters}{a named list or environment}

\item{time_step}{the size of the time step}
}
\value{
a vector of bites on each human
}
\description{
This is an exact algorithm that first samples the number of infectious bites
produced by mosquitoes for each patch from a Poisson distribution and then apportions
them to humans in each patch by Multinomial sampling.
}
\section{input}{

The input \code{parameters} may be a named list or environment with the following
elements
 * a: human biting rate (by patch)
 * Z: number of infectious mosquitoes (by patch)
 * rho: biting weights of all humans
 * loc: locations of all humans
}

