% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_loop.R
\name{step_mainloop}
\alias{step_mainloop}
\title{Main loop for a MASH simulation to run a discrete time step.}
\usage{
step_mainloop(modules)
}
\arguments{
\item{modules}{A list of modules, named `location`, `bloodmeal`,
    and `health`.

This function is responsible for calling the modules in the correct
order and for passing the results of one module to others.
We call the results of a module's time step a path because it is
the route that the simulation took over the time step.
It assumes that each module implements S3 methods for the
\code{\link{mash_step}}, which simulates the next discrete time step.
Each module will also implement S3 methods that extract results
of the time step so that they can be passed, by this function, to
the modules that need them as inputs.}
}
\description{
Main loop for a MASH simulation to run a discrete time step.
}
\seealso{
\code{\link{location_path}}, \code{\link{person_path}},
    \code{\link{human_disease_path}}, \code{\link{infects_mosquito_path}},
    \code{\link{human_infection_path}}
}
