% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forced_si.R
\name{step_si_module}
\alias{step_si_module}
\title{Takes one time step of the discrete time step.}
\usage{
step_si_module(simulation, bites)
}
\arguments{
\item{simulation}{A forced-SI model.}

\item{bites}{A list of bites for each person. Each item in the list is a vector
of times when an infectious bite happens.}
}
\value{
a simulation object
}
\description{
Takes one time step of the discrete time step.
}
\examples{
\dontrun{
step_si_module(simulation, forced_si_create_bites(100, 1/20, current_time, 14))
}
}
