% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strata_dat}
\alias{strata_dat}
\title{Strata: time series w/ meta data}
\format{data frame with 29838 rows and 9 columns}
\usage{
strata_dat
}
\description{
Strata: time series w/ meta data
}
\keyword{datasets}
