% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strata_riws}
\alias{strata_riws}
\title{Strata level RIWs}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 68 rows and 5 columns.}
\usage{
strata_riws
}
\description{
Strata level RIWs
}
\keyword{datasets}
