% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forced_si.R
\name{trajectory_si_module}
\alias{trajectory_si_module}
\title{Extract the trajectory from the simulation}
\usage{
trajectory_si_module(simulation)
}
\arguments{
\item{simulation}{a forced-SI model.}
}
\value{
a list of trajectory entries.
}
\description{
Extract the trajectory from the simulation
}
