% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans}
\alias{trans}
\title{Data: Transportation > Motor fuel}
\format{An object of class \code{data.frame} with 5001 rows and 10 columns.}
\usage{
trans
}
\description{
Indices for Motor Fuel Expenditure Class
}
\keyword{datasets}
