% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{update_individual}
\alias{update_individual}
\title{Update one individual}
\usage{
update_individual(state, transitions, observe)
}
\arguments{
\item{state}{a list of their traits and times}

\item{transitions}{a list of transitions for enabling and firing times}

\item{observe}{A function that records the transition.}
}
\value{
a list with the individual a trajectory entry, and a new time
}
\description{
Update one individual
}
