% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Util_PfSI_State}
\alias{Util_PfSI_State}
\title{Utility: Discretize PfSI Event Output to Daily Jump Process}
\usage{
Util_PfSI_State(out)
}
\arguments{
\item{out}{raw output of PfSI in \code{\link{data.frame}} format}
}
\description{
Take output of PfSI human infection logging and return a day by day state space data frame.
}
