% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{when}
\alias{when}
\title{Find when this transition will fire.}
\usage{
when(transition, individual, when, variables = NULL)
}
\arguments{
\item{transition}{A list of transitions}

\item{individual}{A list with individual properties}

\item{when}{A time}

\item{variables}{Any global simulation state}
}
\value{
numeric time relative to now, when it would fire
}
\description{
This function is for testing the transitions on data for a
single individual. It's a helper function. The running simulation
doesn't call this function.
}
\seealso{
\code{\link{is_enabled}}, \code{\link{fire}}
}
