% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_density.R
\name{assign_levels_to_bites}
\alias{assign_levels_to_bites}
\title{Assign bite levels for bites of a single human.}
\usage{
assign_levels_to_bites(health_rec, bite_cnt, time_cols, level_cols, params)
}
\arguments{
\item{health_rec}{a single row of the health data table.

health_rec <- health_dt\link{health_dt$ID == h_idx,}
time_cols <- grep("Time", names(health_rec))
level_cols <- c(time_cols\link{1} - 1, time_cols + 1)}
}
\description{
Assign bite levels for bites of a single human.
}
