% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_linear.R
\name{bloodmeal_process}
\alias{bloodmeal_process}
\title{Assign bites to people, the top-level function.}
\usage{
bloodmeal_process(health_dt, movement_dt, bites_dt)
}
\arguments{
\item{health_dt}{Health events from the protocol.}

\item{movement_dt}{Movement events from the protocol.}

\item{bites_dt}{Bite events from the protocol.

infect_human <- outcome_dt\link{Bite > 0.0}
infect_mosquito <- outcome_dt\link{(Bite == 0.0) & (Level > 0.0)}}
}
\description{
Assign bites to people, the top-level function.
}
