% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{build_internal_parameters}
\alias{build_internal_parameters}
\title{Turn external parameters into mosquito internal parameters.}
\usage{
build_internal_parameters(parameters)
}
\arguments{
\item{The}{external parameters.}
}
\description{
Turn external parameters into mosquito internal parameters.
}
