% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{continuous_simulation}
\alias{continuous_simulation}
\title{Create a continuous-time simulation}
\usage{
continuous_simulation(individuals, transitions, observer, variables = NULL)
}
\arguments{
\item{individuals}{a data.table with a row for each individual}

\item{transitions}{Each transition is a list of enabling rule, enabling time, and firing
functions.}

\item{observer}{A function that examines each transition and stores information about it.
For example, see \code{\link{observe_continuous}}}

\item{variables}{An environment containing global variables for this simulation.
This can be NULL.}
}
\value{
A simulation object, as a list.
}
\description{
Create a continuous-time simulation
}
