% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{continuous_step}
\alias{continuous_step}
\title{Run until a specified time.}
\usage{
continuous_step(individuals, is_enabled, when, fire)
}
\arguments{
\item{individuals}{a data frame of individuals, including columns for times.}

\item{is_enabled}{a list of enabling functions.}

\item{when}{a list of functions that return firing times when enabled}

\item{fire}{a list of functions that say when to fire.}
}
\value{
a trajectory
}
\description{
Run until a specified time.
}
