% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human-PfSI.R
\name{human_pfsi_conpars}
\alias{human_pfsi_conpars}
\title{Human-PfSI: Individual Constructor Parameters}
\usage{
human_pfsi_conpars(id, home_patch_id, trip_duration, trip_frequency,
  bweight, age, inf, chx)
}
\arguments{
\item{id}{integer id of this person}

\item{home_patch_id}{the id of their home patch}

\item{trip_duration}{numeric value; average durarion of trips away from home}

\item{trip_frequency}{numeric value; average rate at which they make trips away from home}

\item{bweight}{numeric value; biting weight to this person (relative to others in a patch; should have mean 1)}

\item{age}{numeric age (not currently used)}

\item{inf}{logical (start simulation infected or not)}

\item{chx}{logical (start simulation with chemoprophylactic protection or not)}
}
\description{
Generates a single named list with parameters to construct a single
'human_pfsi' object (see inst/include/Human-PfSI.hpp for constructor parameters).
Be aware that all 'id' related fields should be considered 0-indexed (especially the home patch).
}
