% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_loop.R
\name{infects_mosquito_path}
\alias{infects_mosquito_path}
\title{Extract all bites that are infectious to a mosquito from a bloodmeal module.}
\usage{
infects_mosquito_path(bloodmeal_module)
}
\description{
Extract all bites that are infectious to a mosquito from a bloodmeal module.
}
\seealso{
\code{\link{infects_human_path}}
}
