% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{initialize_times}
\alias{initialize_times}
\title{Set up firing times given an initial state.}
\usage{
initialize_times(individuals, transitions, curtime)
}
\arguments{
\item{individuals}{a data frame of individuals, including columns for times.}

\item{transitions}{a list of transitions for enabling and firing times}
}
\value{
a data frame of individuals, with times filled in
}
\description{
Set up firing times given an initial state.
}
