% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\docType{package}
\name{macro}
\alias{macro}
\title{macro: Simulation for vector-borne disease.}
\description{
This package is a set of tools for modular simulation.
}
\section{The four types of modules}{


Every simulation has four module types: a human module, mosquito module,
location module for human location, and a bloodmeal that simulates bites.
}

\section{Available modules}{


\itemize{
\item \code{mosquito_rm} Mosquito simulation using a Ross-MacDonald model.
\item \code{forced_si} Human module.
\item \code{bloodmeal_linear}
\item \code{bloodmeal_density}
\item \code{location_simpletrip}
\item \code{location_flux}
}
}

