% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{make_tosolve}
\alias{make_tosolve}
\title{Find fraction of mosquitoes bitten each day, given fraction infected.}
\usage{
make_tosolve(parameters, location_idx)
}
\arguments{
\item{parameters}{The input module parameters.}
}
\value{
A function to use for optimization.
If we know how many mosquitoes should be infected at steady state,
find the number of bites that must be happening. This has to figure
out how to allocated the infected fraction among Y and Z and invert
the relationship between biting and Y and Z.
}
\description{
Find fraction of mosquitoes bitten each day, given fraction infected.
}
