% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_simpletrip.R
\name{mash_step.simple_trip}
\alias{mash_step.simple_trip}
\title{Takes one time step of the Simple Trip model}
\usage{
\method{mash_step}{simple_trip}(simulation, duration_days, health_path)
}
\arguments{
\item{simulation}{A simple trip model (most likely made via \code{\link[macro]{simple_trip_module}})}

\item{duration_days}{how long to step the simulation for}

\item{health_path}{The history of health states for each person.}
}
\description{
Takes one time step of the Simple Trip model
}
