% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mosquito_rm_dynamics}
\alias{mosquito_rm_dynamics}
\title{Dynamics of Ross-Macdonald mosquitoes.}
\usage{
mosquito_rm_dynamics(state, parameters, bites, aquatic = mosquito_rm_aquatic)
}
\arguments{
\item{state}{The internal state.}

\item{parameters}{The internal parameters.}

\item{bites}{Net infectiousness of humans to mosquitoes expressed
as a number of infectious bites in each location. The units on bites
must match the M units, which can be either a fraction or a count.}

\item{aquatic}{The function to call to get the emergence.}
}
\value{
a list with five values: the vectors \code{M}, \code{Y}, \code{Z}, and the
parameters \code{a} and \code{simulation_day}.

This is a single-day dynamics step for a discrete-time analog
of the Ross-Macdonald model. The aquatic lifecycle is stochastic, but
the rest is deterministic.
}
\description{
Dynamics of Ross-Macdonald mosquitoes.
}
