% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{mosquito_rm_dynamics}
\alias{mosquito_rm_dynamics}
\title{Dynamics of R-M mosquitoes.}
\usage{
mosquito_rm_dynamics(state, parameters, kappa, aquatic = mosquito_rm_aquatic)
}
\arguments{
\item{state}{The internal state.}

\item{parameters}{The internal parameters.}

\item{kappa}{Net infectiousness of humans to mosquitoes, the probability
a mosquito becomes infected after feeding on a human. Vector per patch.}

\item{aquatic}{The function to call to get the emergence.}
}
\description{
Dynamics of R-M mosquitoes.
}
