% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msource.R
\name{msource}
\alias{msource}
\title{Preprocess and Source}
\usage{
msource(pth)
}
\arguments{
\item{pth}{The path to the R program to source.}
}
\value{
The path of the sourced program.
}
\description{
The \code{msource} function runs the macro preprocessor
and then executes the program normally.
}
