% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msource.R
\name{msource}
\alias{msource}
\title{Pre-process and Source}
\usage{
msource(
  pth = Sys.path(),
  file_out = NULL,
  envir = globalenv(),
  exec = TRUE,
  debug = FALSE,
  debug_out = NULL,
  ...
)
}
\arguments{
\item{pth}{The path to the R program to source.}

\item{file_out}{If you want to save or view the generated code
from the \code{msource} function, supply a full path and
file name on this parameter. Default is NULL. When NULL,
the function will create a temp file for the generated code.}

\item{envir}{The environment to be used for program execution.
Default is the global environment.}

\item{exec}{Whether or not to execute the output file after pre-processing.
Default is TRUE}

\item{debug}{If TRUE, prints lines to the console as they are processed.
This information case be useful for debugging macro code.
Default is FALSE.  See the \code{vignette("macro-debug"} for more information
on debugging.}

\item{debug_out}{A path to a file to be used for debugging.  If a path
is supplied, debug output will be written to the file instead of the
console. Default is NULL.}

\item{...}{Follow-on parameters to the \code{source} function. See
the \code{\link{source}} function for additional information.}
}
\value{
The results of the \code{source()} function, invisibly.  The path
of the resolved output file is also included under the "$output" list item.
}
\description{
The \code{msource} function runs the macro pre-processor
and then executes the program normally.
}
\details{
Additional details...
}
\section{Macro Commands}{

Below is the complete list of available macro commands:
\itemize{
\item{\strong{#\%<comment>}: A macro comment.}
\item{\strong{#\%include '<path>'}: Inserts code from included file as text
into current program.}
\item{\strong{#\%if (<condition>)}: Begins a macro conditional block.}
\item{\strong{#\%elseif (<condition>)}: Defines a subsequent conditional block.}
\item{\strong{#\%else}: Identifies the default behavior in a condition.}
\item{\strong{#\%end}: Ends a macro condition.}
\item{\strong{\%sysfunc(<expression>)}: Executes an R expression as part of
a macro command.}
}
}

