% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msource.R
\name{msource}
\alias{msource}
\title{Preprocess and Source}
\usage{
msource(pth, file_out = NULL)
}
\arguments{
\item{pth}{The path to the R program to source.}

\item{file_out}{If you want to save or view the generated code
from the \code{msource} function, supply a full path and
file name on this parameter.}
}
\value{
The path of the sourced program.
}
\description{
The \code{msource} function runs the macro preprocessor
and then executes the program normally.
}
