% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\name{observe_bloodmeal_mosquito}
\alias{observe_bloodmeal_mosquito}
\title{Observe the bloodmeal bites to mosquitoes.}
\usage{
observe_bloodmeal_mosquito(observer, bloodmeal_path, step)
}
\description{
Observe the bloodmeal bites to mosquitoes.
}
