% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{runMSourceDebug}
\alias{runMSourceDebug}
\title{Addin Function to Run msource() in Debug Mode}
\usage{
runMSourceDebug()
}
\value{
The results of \code{msource}, invisibly.
}
\description{
This function is exposed to the addin menu to run \code{msource()}
interactively in debug mode. The function will run either the
currently selected code, or the currently active program.
On the call to \code{msource}, it sets the "debug" and "symbolgen"
parameters to TRUE.  The function also sets the "envir" parameter
to the global environment and the "clear" parameter to FALSE
to enhance the user experience.  The function
takes no parameters and is only used by the addin menu.
}
\examples{
\donttest{
# Called from addin menu
runMSourceDebug()
}
}
