% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal_daily.R
\name{sample_bites}
\alias{sample_bites}
\title{Determines an integral numbers of bites at each location.}
\usage{
sample_bites(travel, mosquitoes, world)
}
\arguments{
\item{travel}{is the location x human matrix of dwell times for each human.}

\item{world}{is parameters for the simulation.}
}
\value{
a location x human matrix of bites.
}
\description{
Determines an integral numbers of bites at each location.
}
