% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_rm.R
\name{shift_with_open_interval}
\alias{shift_with_open_interval}
\title{Makes a shift matrix that accumulates the oldest.}
\usage{
shift_with_open_interval(N)
}
\arguments{
\item{N}{The size of the square matrix.}
}
\description{
Makes a shift matrix that accumulates the oldest.
}
