% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_step.R
\name{update_individual}
\alias{update_individual}
\title{Update one individual}
\usage{
update_individual(next_person, is_enabled, when, fire)
}
\arguments{
\item{next_person}{a list of their traits and times}

\item{is_enabled}{a list of functions to determine enabling}

\item{when}{a list of functions to determine when to fire}

\item{fire}{a list of functions to determine firing}
}
\value{
a list with the individual a trajectory entry, and a new time
}
\description{
Update one individual
}
